1_foundations/requirements.txt
requests
python-dotenv
gradio
pypdf
openai
resend